Allow config location to be customized from CLI (#66)
* Allow config location to be customized from CLI

Config file should be stored in a persistent storage as it is stores the
app status and configuration. It should be independent from container
life cycle and mounted into app container. This commit allows config
location to be changed through CLI. So that when the persistent storage
is mounted, the location can be pointed to the mounted config file.

Also some updates for best practices:
* extracted version numbers in Dockerfile into a environment variable
* Added mattermost/bin to PATH environment variable

* Add customizable environment variables to readme